+++ data-raw/packages.R
@@ -50,7 +50,8 @@ paletteer_packages <- tibble::tribble(
   "colRoz", "jacintak/colRoz", "0.2.2", FALSE, NA,
   "unikn", "hneth/unikn", "0.2.0.9003", TRUE, "0.2.0",
   "tvthemes", "Ryo-N7/tvthemes", "1.1.0", TRUE, "1.1.0",
-  "PNWColors", "jakelawlor/PNWColors", "0.1.0", FALSE, NA
+  "PNWColors", "jakelawlor/PNWColors", "0.1.0", FALSE, NA,
+  "palettesForR", "frareb/palettesForR", "0.1.2", TRUE, "0.1.2"
 ) %>%
   arrange(Name)